backend/requirements.txt
django==4.2.9
djangorestframework==3.14.0
channels==4.0.0
# channels_redis
# celery==5.3.6
channels_redis==4.2.0
celery[redis]==5.3.6
# uvicorn==0.27.0
uvicorn[standard]==0.27.0